Split WinMDWriter for better separation of concerns
Extract finalization logic from WinMDWriter.cs into WinMDWriter.Finalization.cs:
- FinalizeGeneration (4-phase post-processing pipeline)
- AddMethodImplFixups / AddMethodImplsForInterface (MethodImpl wiring)
- FindMatchingMethod / FindExplicitMethodImpl (method resolution)
- IsProjectionEquivalent (.NET ↔ WinRT type equivalence)
- AddOverloadAttributesForType / AddOverloadAttribute (WinRT overload naming)

The main WinMDWriter.cs now only contains state, constructor, ProcessType,
SafeResolve, and Write — matching the interop generator's pattern of keeping
the orchestration file slim and splitting implementation by concern.
